polish_pass
