+++ trilogy/parsing/v2/lark_backend.py
@@ -5,7 +5,11 @@ from os.path import dirname, join
 from re import IGNORECASE
 from typing import TYPE_CHECKING, Any
 
-from trilogy.parsing.v2.errors import create_generic_syntax_error, create_syntax_error
+from trilogy.parsing.v2.errors import (
+    create_generic_syntax_error,
+    create_syntax_error,
+    detect_subselect,
+)
 from trilogy.parsing.v2.syntax import SyntaxDocument, syntax_document_from_parser
 from trilogy.utility import safe_open
 
@@ -127,6 +131,10 @@ def _handle_unexpected_token(e: "UnexpectedToken", text: str) -> None:
     by_pos = _detect_unparenthesized_by_expr_lark(text, pos)
     if by_pos is not None:
         raise create_syntax_error(211, by_pos, text)
+
+    sub_pos = detect_subselect(text, pos)
+    if sub_pos is not None:
+        raise create_syntax_error(102, sub_pos, text)
 
     if last_token and e.token.type == "$END":
         try:
